do not fetch headers when renaming all or none
Renaming all files or no files is now instant, because it doesn't depend
on whether the filename exists in the remote header.

Add error handling for when we make requests and parse xml. Catch
specific errors so we don't accidentally swallow a keyboard interrupt.

Catch OSError when running main(), in case the input or output file
can't be opened.

Use fstrings in place of concatenation